Extract sound effects enum into its own file
Avoids depending on all of `effects` in places where we only need the
enum.
