+++ Source/monstdat.h
@@ -5,7 +5,9 @@
  */
 #pragma once
 
+#include <cstddef>
 #include <cstdint>
+#include <string>
 #include <vector>
 
 #include "cursor.h"
